+++ crates/api-desc/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Minor
 
+- Add `ec` module with ECDH
 - Add `hash` module with SHA-256 algorithm
 - Support in-place AES-256-GCM by accepting null pointers