+++ crates/prelude/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Minor
 
+- Add ECDH in `crypto::ec`
 - Add SHA-256 in `crypto::hash`
 - Add `sync` module for mutex and atomics support
 - Add in-place variants for AES-256-GCM